0309 작업 모바일 메인페이지와 동일하게 제작 (UI)
