+++ src/main/frontend/src/App.scss
@@ -1,97 +1,95 @@
-@import "./color.scss";
-
 // FIXME: Tailwind CSS 반응형 적용하기
-// FIXME: dotDiv -> Tailwind CSS 방식으로 적용하기
-.dotDiv {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-}
+// FIXME: dotDiv -> Tailwind CSS 방식으로 적용하기 // 완료
+// .dotDiv {
+//   display: flex;
+//   justify-content: center;
+//   align-items: center;
+//   position: absolute;
+// }
 /* 모바일 화면에 대한 스타일 */
-@media (max-width: 768px) {
-  body {
-    margin: 0 5px;
-  }
-  // FIXME: dot, dotCgv들 Tailwind로 적용하기
-  .dot {
-    width: 6px;
-    height: 6px;
-    margin-left: 2px;
-  }
-  .dotCgv {
-    background-color: $cgv;
-  }
-  .dotMega {
-    background-color: $mega;
-  }
-  .dotLotte {
-    background-color: $lotte;
-  }
-  /* .contents {
+// @media (max-width: 768px) {
+//   body {
+//     margin: 0 5px;
+//   }
+// FIXME: dot, dotCgv들 Tailwind로 적용하기 // 완료
+// .dot {
+//   width: 6px;
+//   height: 6px;
+//   margin-left: 2px;
+// }
+// .dotCgv {
+//   background-color: $cgv;
+// }
+// .dotMega {
+//   background-color: $mega;
+// }
+// .dotLotte {
+//   background-color: $lotte;
+// }
+/* .contents {
     margin-top: 16px;
     margin-bottom: 24px;
-  } */
-  .dotDiv {
-    transform: translate(-8%, 200%);
-  }
-}
+//   } */
+//   .dotDiv {
+//     transform: translate(-8%, 200%);
+//   }
+// }
 
 /* 태블릿 화면에 대한 스타일 */
-@media (min-width: 769px) and (max-width: 1366px) {
-  body {
-    margin: 0 30px;
-  }
-  .dot {
-    width: 8px;
-    height: 8px;
-    margin-left: 2px;
-  }
-  .dotCgv {
-    background-color: $cgv;
-  }
-  .dotMega {
-    background-color: $mega;
-  }
-  .dotLotte {
-    background-color: $lotte;
-  }
-  /* .contents {
+// @media (min-width: 769px) and (max-width: 1366px) {
+//   body {
+//     margin: 0 30px;
+//   }
+//   .dot {
+//     width: 8px;
+//     height: 8px;
+//     margin-left: 2px;
+//   }
+//   .dotCgv {
+//     background-color: $cgv;
+//   }
+//   .dotMega {
+//     background-color: $mega;
+//   }
+//   .dotLotte {
+//     background-color: $lotte;
+//   }
+/* .contents {
     margin-top: 25px;
     margin-bottom: 35px;
   } */
-  .dotDiv {
-    transform: translate(-8%, 150%);
-  }
-}
+//   .dotDiv {
+//     transform: translate(-8%, 150%);
+//   }
+// }
 
 /* 웹 화면에 대한 스타일 */
-@media (min-width: 1367px) {
-  body {
-    margin: 0 113px;
-  }
-  .dot {
-    width: 8px;
-    height: 8px;
-    margin-left: 2px;
-  }
-  .dotCgv {
-    background-color: $cgv;
-  }
-  .dotMega {
-    background-color: $mega;
-  }
-  .dotLotte {
-    background-color: $lotte;
-  }
-  /* .contents {
-    margin-top: 25px;
-    margin-bottom: 35px;
-  } */
-  .dotDiv {
-    transform: translate(-8%, 150%);
-  }
-}
+// @media (min-width: 1367px) {
+//   body {
+//     margin: 0 113px;
+//   }
+//   .dot {
+//     width: 8px;
+//     height: 8px;
+//     margin-left: 2px;
+//   }
+//   .dotCgv {
+//     background-color: $cgv;
+//   }
+//   .dotMega {
+//     background-color: $mega;
+//   }
+//   .dotLotte {
+//     background-color: $lotte;
+//   }
+//   /* .contents {
+//     margin-top: 25px;
+//     margin-bottom: 35px;
+//   } */
+//   .dotDiv {
+//     transform: translate(-8%, 150%);
+//   }
+// }
 
 /* .absoluteDiv {
   display: flex;
