--- src/main/frontend/src/color.scss
@@ -1,4 +1,0 @@
-// FIXME: Tailwind로 선언 후 삭제
-$cgv: #e42313;
-$mega: #362462;
-$lotte: #ee870e;
