--- src/main/frontend/src/components/Footer/footer.scss
@@ -1,18 +1,0 @@
-// FIXME: 삭제
-footer {
-  padding: 10px;
-  text-align: center; /* 가운데 정렬 */
-
-  bottom: 0;
-  width: 100%;
-  height: 40px;
-  border-top: 1px solid #8c8d96;
-  color: #8c8d96;
-}
-
-footer p {
-  margin: 0;
-  font-size: 14px;
-  color: #666;
-  line-height: 40px; /* 세로 가운데 정렬 시도 */
-}
