--- src/main/frontend/src/components/Header/header.scss
@@ -1,8 +1,0 @@
-// FIXME: 삭제
-header {
-  border-bottom: 1px solid#8c8d96;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 80px;
-}
