--- src/main/frontend/src/components/Popup/popup.scss
@@ -1,10 +1,0 @@
-// FIXME: Tailwind방식 적용 후 삭제
-.popup {
-  z-index: 4;
-  background-color: white;
-  border-radius: 8px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1); /* 그림자 효과 */
-  align-items: center;
-}
-.popup-content {
-}
