--- src/main/frontend/src/components/Search/Searchpage.scss
@@ -1,78 +1,0 @@
-// FIXME: Tailwind 방식으로 적용하고 삭제
-
-/* Searchpage.css */
-.searchpage-modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  width: 100%;
-  height: 100%;
-  z-index: 1000; /* 다른 요소 위에 레이어를 만들기 위한 z-index 값 */
-}
-
-.search-contain {
-  width: 100%;
-  display: flex;
-  z-index: 1001; /* 모달 내부 요소는 모달 위에 레이어를 만들기 위한 z-index 값 */
-}
-
-.search-input {
-  visibility: visible;
-  border: none;
-  background: #f5f5f5;
-  width: 106%;
-  cursor: text;
-  color: inherit;
-  padding: 10px;
-  margin: 17px 70px 0 15px;
-  border-radius: 6px;
-}
-
-.btn-search,
-.btn-cancel {
-  font-size: 16px;
-  position: absolute;
-  top: 24px;
-  right: 35px;
-  border: none;
-  background-color: transparent;
-  font-weight: bold;
-}
-
-.recent-search {
-  margin: 40px 30px 0 24px;
-  display: flex;
-  position: relative;
-  p {
-    font-size: 12px;
-  }
-}
-
-.btn-recent-delete {
-  font-size: 12px;
-  position: absolute;
-  top: 10px;
-  right: 0px;
-  border: none;
-  background-color: transparent;
-}
-
-.recent-list {
-  position: absolute;
-  top: 30px;
-  left: 15px;
-}
-
-.recent-list li {
-  all: unset;
-  display: list-item;
-  list-style: initial;
-}
-
-.btn-delete {
-  border: none;
-  background-color: transparent;
-  justify-content: flex-end;
-}
