--- src/main/frontend/src/components/Searchicon/search.scss
@@ -1,23 +1,0 @@
-// FIXME: Search에 Tailwind 방식으로 적용 후 파일 삭제
-
-.searchicon {
-  width: 16px;
-  height: 16px;
-  margin-left: 6px;
-  margin-right: 6px;
-}
-
-/* 모바일 화면에 대한 스타일 */
-@media (max-width: 768px) {
-}
-
-/* 태블릿 화면에 대한 스타일 */
-@media (min-width: 769px) and (max-width: 1366px) {
-}
-
-/* 웹 화면에 대한 스타일 */
-@media (min-width: 1367px) {
-  .searchicon {
-    display: none;
-  }
-}
