+++ src/main/frontend/src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.scss";
-import "./tailwind.css";
 import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
