+++ src/main/frontend/src/index.scss
@@ -4,10 +4,21 @@
 @tailwind components;
 @tailwind utilities;
 
-// FIXME: Pretendard로 웹폰트 적용하기
+// FIXME: Pretendard로 웹폰트 적용하기 // 완료
+
+@font-face {
+  font-family: "Pretendard-Regular";
+  src: url("./fonts/PretendardVariable.ttf") format("truetype");
+}
 
 // FIXME: 폰트 사이즈 규격에 대한 커스텀 클래스 설정하는것 추천
 @layer components {
+  .font-10 {
+    font-size: 10px;
+    line-height: 16px;
+    letter-spacing: -0.005em;
+  }
+
   .font-12 {
     font-size: 12px;
     line-height: 16px;
@@ -19,18 +30,33 @@
     line-height: 20px;
     letter-spacing: -0.01em;
   }
+  .font-14 {
+    font-size: 14px;
+    line-height: 20px;
+    letter-spacing: -0.01em;
+  }
 }
 
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+  font-family:
+    "Pretendard-Regular",
+    -apple-system,
+    BlinkMacSystemFont,
+    "Segoe UI",
+    "Roboto",
+    "Oxygen",
+    "Ubuntu",
+    "Cantarell",
+    "Fira Sans",
+    "Droid Sans",
+    "Helvetica Neue",
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+  font-family: "Pretendard-Regular", source-code-pro, Menlo, Monaco, Consolas,
+    "Courier New", monospace;
 }
