--- src/main/frontend/src/styles/base/_reset.scss
@@ -1,50 +1,0 @@
-* {
-  margin: 0;
-  font-family: $font-main;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  box-sizing: border-box;
-}
-
-html {
-  font-family: $font-main;
-  font-size: $font-size-16;
-}
-
-body {
-  font-family: $font-main;
-  color: $primary;
-}
-
-h1 {
-  margin: 0;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-button,
-input,
-select,
-textarea {
-  background-color: transparent;
-  border: 0;
-
-  &:focus {
-    outline: none;
-    box-shadow: none;
-  }
-}
-
-a,
-button {
-  cursor: pointer;
-}
-
-ol,
-li {
-  padding-left: 0;
-  list-style: none;
-}
