--- src/main/frontend/src/styles/constants/_breakpoints.scss
@@ -1,2 +1,0 @@
-$md-breakpoint: 768px;
-$lg-breakpoint: 1200px;
