--- src/main/frontend/src/styles/constants/_colors.scss
@@ -1,11 +1,0 @@
-$black: #000;
-$primary: #323232;
-$secondary: #8c8d96;
-$tertiary: #b2b3b9;
-$border: #d1d1d1;
-$background: #f5f5f5;
-$white: #fff;
-
-$red: #e42313;
-$orange: #ee870e;
-$purple: #362462;
