--- src/main/frontend/src/styles/constants/_grids.scss
@@ -1,14 +1,0 @@
-$gutter: 20px;
-
-$sm-columns: 4;
-$sm-margin: 5px;
-
-$md-columns: 12;
-$md-margin: 30px;
-$md-max-container: 960px + $md-margin * 2;
-
-$lg-columns: 12;
-$lg-unit: 75px;
-$lg-max-container: ($lg-unit + $gutter) * $lg-columns;
-
-$grid-margin: 15px;
