--- src/main/frontend/src/styles/constants/_typography.scss
@@ -1,26 +1,0 @@
-$font-main: "Noto Sans KR", sans-serif;
-$font-price: "Tahoma", sans-serif;
-
-$font-size-12: 12px;
-$line-height-12: 16px;
-$letter-spacing-12: -0.005em;
-
-$font-size-13: 13px;
-$line-height-13: 20px;
-$letter-spacing-13: -0.01em;
-
-$font-size-14: 14px;
-$line-height-14: 24px;
-$letter-spacing-14: -0.01em;
-
-$font-size-16: 16px;
-$line-height-16: 24px;
-$letter-spacing-16: -0.01em;
-
-$font-size-18: 18px;
-$line-height-18: 28px;
-$letter-spacing-18: -0.02em;
-
-$font-size-24: 24px;
-$line-height-24: 34px;
-$letter-spacing-24: -0.01em;
