--- src/main/frontend/src/styles/layouts/_grids.scss
@@ -1,42 +1,0 @@
-.container {
-  width: 100%;
-  padding: 0 $sm-margin;
-  margin: 0 auto;
-
-  .row {
-    display: flex;
-    flex-wrap: wrap;
-  }
-
-  [class^="col-"] {
-    padding: 0 $gutter / 2;
-  }
-
-  @for $i from 1 through $sm-columns {
-    .col-sm-#{$i} {
-      width: percentage($i / $sm-columns);
-    }
-  }
-
-  @include responsive(T) {
-    max-width: $md-max-container;
-    padding: 0 $md-margin;
-
-    @for $i from 1 through $md-columns {
-      .col-md-#{$i} {
-        width: percentage($i / $md-columns);
-      }
-    }
-  }
-
-  @include responsive(D) {
-    max-width: $lg-max-container;
-    padding: 0;
-
-    @for $i from 1 through $lg-columns {
-      .col-lg-#{$i} {
-        width: ($lg-unit + $gutter) * $i;
-      }
-    }
-  }
-}
