--- src/main/frontend/src/styles/main.scss
@@ -1,19 +1,0 @@
-// FIXME: Sass 사용 안하니까 styles 폴더 삭제하기
-
-@import "./constants/colors";
-@import "./constants/typography";
-@import "./constants/grids";
-@import "./constants/breakpoints";
-
-@import "./base/normalisze";
-@import "./base/reset";
-@import "./base/globals";
-
-@import "./mixins/text-style";
-@import "./mixins/flexbox";
-@import "./mixins/positions";
-@import "./mixins/responsive";
-
-@import "./modules/tag";
-
-@import "./layouts/grids";
