--- src/main/frontend/src/styles/mixins/_flexbox.scss
@@ -1,31 +1,0 @@
-$flex-map: (
-  start: flex-start,
-  end: flex-end,
-  between: space-between,
-  around: space-around,
-  stretch: stretch,
-  center: center,
-);
-
-@function _get-flex-value($key) {
-  @return map-get($flex-map, $key);
-}
-
-@mixin flexbox($jc: center, $ai: center) {
-  display: flex;
-  justify-content: _get-flex-value($jc);
-  align-items: _get-flex-value($ai);
-}
-
-@mixin inline-flexbox($jc: center, $ai: center) {
-  display: inline-flex;
-  justify-content: _get-flex-value($jc);
-  align-items: _get-flex-value($ai);
-}
-
-@mixin column-flexbox($jc: center, $ai: center) {
-  display: flex;
-  flex-direction: column;
-  justify-content: _get-flex-value($jc);
-  align-items: _get-flex-value($ai);
-}
