--- src/main/frontend/src/styles/mixins/_positions.scss
@@ -1,24 +1,0 @@
-@mixin pos-center-x($type: absolute) {
-  @if ($type == fixed or $type == absolute) {
-    position: $type;
-    left: 50%;
-    transform: translateX(-50%);
-  }
-}
-
-@mixin pos-center-y($type: absolute) {
-  @if ($type == fixed or $type == absolute) {
-    position: $type;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-}
-
-@mixin pos-center($type: absolute) {
-  @if ($type == fixed or $type == absolute) {
-    position: $type;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-}
