--- src/main/frontend/src/styles/mixins/_responsive.scss
@@ -1,19 +1,0 @@
-@mixin responsive($screen) {
-  @if ($screen == "T") {
-    @media screen and (min-width: $md-breakpoint) {
-      @content;
-    }
-  }
-
-  @if ($screen == "D") {
-    @media screen and (min-width: $lg-breakpoint) {
-      @content;
-    }
-  }
-}
-
-@mixin mobile-only() {
-  @media screen and (max-width: $md-breakpoint - 1px) {
-    @content;
-  }
-}
