--- src/main/frontend/src/styles/mixins/_text-style.scss
@@ -1,60 +1,0 @@
-@mixin text-style-12() {
-  font-size: $font-size-12;
-  line-height: $line-height-12;
-  letter-spacing: $letter-spacing-12;
-}
-
-@mixin text-style-13() {
-  font-size: $font-size-13;
-  line-height: $line-height-13;
-  letter-spacing: $letter-spacing-13;
-}
-
-@mixin text-style-14() {
-  font-size: $font-size-14;
-  line-height: $line-height-14;
-  letter-spacing: $letter-spacing-14;
-}
-
-@mixin text-style-16() {
-  font-size: $font-size-16;
-  line-height: $line-height-16;
-  letter-spacing: $letter-spacing-16;
-}
-
-@mixin text-style-18() {
-  font-size: $font-size-18;
-  line-height: $line-height-18;
-  letter-spacing: $letter-spacing-18;
-}
-
-@mixin text-style-24() {
-  font-size: $font-size-24;
-  line-height: $line-height-24;
-  letter-spacing: $letter-spacing-24;
-}
-
-@mixin text-style($size, $color: false) {
-  @if ($size == 12) {
-    @include text-style-12;
-  }
-  @if ($size == 13) {
-    @include text-style-13;
-  }
-  @if ($size == 14) {
-    @include text-style-14;
-  }
-  @if ($size == 16) {
-    @include text-style-16;
-  }
-  @if ($size == 18) {
-    @include text-style-18;
-  }
-  @if ($size == 24) {
-    @include text-style-24;
-  }
-
-  @if (type-of($color) == color) {
-    color: $color;
-  }
-}
