--- src/main/frontend/src/tailwind.css
@@ -1,4 +1,0 @@
-/* FIXME: 파일 삭제*/
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
